+++ Cargo.toml
@@ -13,6 +13,4 @@ log = "0.4"
 serenity = "0.8.0"
 serde_json = "1.0.40"
 reqwest = { version = "0.9.19", default-features = false, features = ["rustls-tls"] }
-
-[dependencies.curl]
-git = "https://github.com/carllerche/curl-rust"
+async-trait = "0.1.26"
