+++ src/listeners/resume.rs
@@ -1,0 +1,6 @@
+use log::info;
+use serenity::{model::prelude::*, prelude::*};
+
+pub fn resume(_: Context, _: ResumedEvent) {
+    info!("Resumed");
+}
